Making Elasticsearch Version injectable. (#3896)
* Making Elasticsearch Version injectable.

This allows us to reuse the mechanism of retrieving the version of the
connected Elasticsearch nodes in other parts of the code besides the
IndexMappingFactory.

* Restructuring and adding tests.

* Making annotation runtime retained and suitable for bindings.

* Adding license headers.

* Rewriting provider to generic service class.

This prevents I/O being done during injection, which could cause to
erratic and intransparent startup behavior in case of errors.

* Correcting check in test.

* Add license header.
