+++ graylog2-server/src/test/java/org/graylog2/indexer/IndexMappingFactoryTest.java
@@ -16,11 +16,8 @@
  */
 package org.graylog2.indexer;
 
-import com.github.zafarkhaja.semver.ParseException;
-import com.google.gson.JsonObject;
-import io.searchbox.client.JestClient;
-import io.searchbox.client.JestResult;
-import io.searchbox.core.Ping;
+import com.github.zafarkhaja.semver.Version;
+import org.graylog2.indexer.cluster.Node;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -30,14 +27,12 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
-import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 public class IndexMappingFactoryTest {
@@ -45,43 +40,20 @@ public class IndexMappingFactoryTest {
     public final MockitoRule mockitoRule = MockitoJUnit.rule();
 
     @Mock
-    private JestClient jestClient;
+    private Node node;
+
     private IndexMappingFactory indexMappingFactory;
 
     @Before
     public void setUp() throws Exception {
-        indexMappingFactory = new IndexMappingFactory(jestClient);
-    }
-
-    @Test
-    public void createIndexMappingFailsIfElasticsearchIsUnavailable() throws Exception {
-        when(jestClient.execute(any(Ping.class))).thenThrow(IOException.class);
-
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
-                .isInstanceOf(ElasticsearchException.class)
-                .hasCauseInstanceOf(IOException.class);
-    }
-
-    @Test
-    public void createIndexMappingFailsIfElasticsearchResponseFailed() throws Exception {
-        final JestResult failedResult = mock(JestResult.class);
-        when(failedResult.isSucceeded()).thenReturn(false);
-        when(failedResult.getJsonObject()).thenReturn(new JsonObject());
-        when(jestClient.execute(any(Ping.class))).thenReturn(failedResult);
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
-                .isInstanceOf(ElasticsearchException.class)
-                .hasMessageStartingWith("Unable to retrieve Elasticsearch version")
-                .hasNoCause();
+        this.indexMappingFactory = new IndexMappingFactory(node);
     }
 
     @Test
     public void createIndexMappingFailsIfElasticsearchVersionIsTooLow() throws Exception {
-        final JestResult jestResult = mock(JestResult.class);
-        when(jestResult.isSucceeded()).thenReturn(true);
-        when(jestResult.getJsonObject()).thenReturn(buildVersionJsonObject("1.7.3"));
-        when(jestClient.execute(any(Ping.class))).thenReturn(jestResult);
+        when(node.getVersion()).thenReturn(Optional.of(Version.valueOf("1.7.3")));
 
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
+        assertThatThrownBy(indexMappingFactory::createIndexMapping)
                 .isInstanceOf(ElasticsearchException.class)
                 .hasMessageStartingWith("Unsupported Elasticsearch version: 1.7.3")
                 .hasNoCause();
@@ -89,12 +61,9 @@ public class IndexMappingFactoryTest {
 
     @Test
     public void createIndexMappingFailsIfElasticsearch2VersionIsTooLow() throws Exception {
-        final JestResult jestResult = mock(JestResult.class);
-        when(jestResult.isSucceeded()).thenReturn(true);
-        when(jestResult.getJsonObject()).thenReturn(buildVersionJsonObject("2.0.0"));
-        when(jestClient.execute(any(Ping.class))).thenReturn(jestResult);
+        when(node.getVersion()).thenReturn(Optional.of(Version.valueOf("2.0.0")));
 
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
+        assertThatThrownBy(indexMappingFactory::createIndexMapping)
                 .isInstanceOf(ElasticsearchException.class)
                 .hasMessageStartingWith("Unsupported Elasticsearch version: 2.0.0")
                 .hasNoCause();
@@ -102,36 +71,12 @@ public class IndexMappingFactoryTest {
 
     @Test
     public void createIndexMappingFailsIfElasticsearchVersionIsTooHigh() throws Exception {
-        final JestResult jestResult = mock(JestResult.class);
-        when(jestResult.isSucceeded()).thenReturn(true);
-        when(jestResult.getJsonObject()).thenReturn(buildVersionJsonObject("6.0.0"));
-        when(jestClient.execute(any(Ping.class))).thenReturn(jestResult);
+        when(node.getVersion()).thenReturn(Optional.of(Version.valueOf("6.0.0")));
 
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
+        assertThatThrownBy(indexMappingFactory::createIndexMapping)
                 .isInstanceOf(ElasticsearchException.class)
                 .hasMessageStartingWith("Unsupported Elasticsearch version: 6.0.0")
                 .hasNoCause();
-    }
-
-    @Test
-    public void createIndexMappingFailsIfElasticsearchVersionIsInvalid() throws Exception {
-        final JestResult jestResult = mock(JestResult.class);
-        when(jestResult.isSucceeded()).thenReturn(true);
-        when(jestResult.getJsonObject()).thenReturn(buildVersionJsonObject("Foobar"));
-        when(jestClient.execute(any(Ping.class))).thenReturn(jestResult);
-
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
-                .isInstanceOf(ElasticsearchException.class)
-                .hasMessageStartingWith("Unable to parse Elasticsearch version: Foobar")
-                .hasCauseInstanceOf(ParseException.class);
-    }
-
-    private static JsonObject buildVersionJsonObject(String foobar) {
-        final JsonObject versionObject = new JsonObject();
-        versionObject.addProperty("number", foobar);
-        final JsonObject jsonObject = new JsonObject();
-        jsonObject.add("version", versionObject);
-        return jsonObject;
     }
 
     @RunWith(Parameterized.class)
@@ -159,7 +104,8 @@ public class IndexMappingFactoryTest {
         private final Class<? extends IndexMapping> expectedMapping;
 
         @Mock
-        private JestClient jestClient;
+        private Node node;
+
         private IndexMappingFactory indexMappingFactory;
 
 
@@ -170,17 +116,12 @@ public class IndexMappingFactoryTest {
 
         @Before
         public void setUp() throws Exception {
-            indexMappingFactory = new IndexMappingFactory(jestClient);
+            when(node.getVersion()).thenReturn(Optional.of(Version.valueOf(this.version)));
+            indexMappingFactory = new IndexMappingFactory(node);
         }
 
         @Test
         public void test() throws Exception {
-            final JestResult jestResult = mock(JestResult.class);
-            when(jestResult.isSucceeded()).thenReturn(true);
-            final JsonObject jsonObject = buildVersionJsonObject(version);
-            when(jestResult.getJsonObject()).thenReturn(jsonObject);
-            when(jestClient.execute(any(Ping.class))).thenReturn(jestResult);
-
             assertThat(indexMappingFactory.createIndexMapping()).isInstanceOf(expectedMapping);
         }
     }
